add factory  handlers to read documents; implement factory functions on all controllers

+++ project1-natours-API/controllers/handlerFactory.js
@@ -1,10 +1,11 @@
 const catchAsync = require('../utils/catchAsync');
 const AppError = require('../utils/appError');
+const APIFeatures = require('../utils/APIFeatures');
 
 exports.deleteOneById = (Model, options = {}) =>
   catchAsync(async (req, res, next) => {
     let doc;
-    if (options.requireOwnership) {
+    if (options.requireOwnership && req.user.role !== 'admin') {
       doc = await Model.findOneAndDelete({ _id: req.params.id, user: req.user.id });
       if (!doc) {
         return next(new AppError('You are not authorized to delete this document', 403));
@@ -67,6 +68,51 @@ exports.createOne = (Model, options = {}) =>
       status: 'success',
       data: {
         data: doc,
+      },
+    });
+  });
+
+exports.getOneById = (Model, populateOptions = []) =>
+  catchAsync(async (req, res, next) => {
+    let query = Model.findById(req.params.id);
+    populateOptions.forEach((option) => {
+      query = query.populate(option);
+    });
+    const doc = await query;
+
+    if (!doc) {
+      return next(new AppError(`No document found with that ID: ${req.params.id}`), 404);
+    }
+    res.status(200).json({
+      status: 'success',
+      data: {
+        data: doc,
+      },
+    });
+  });
+
+exports.getAll = (Model, populateOptions = []) =>
+  catchAsync(async (req, res, next) => {
+    // To allow for nested GET reviews on Tours
+    let criteria;
+    if (req.params.tourId) criteria = { tour: req.params.tourId };
+    // Query
+    const features = new APIFeatures(Model.find(criteria), req.query)
+      .filter()
+      .sort()
+      .limit()
+      .pagination();
+    // eslint-disable-next-line prefer-destructuring
+    let query = features.query;
+    populateOptions.forEach((option) => {
+      query = query.populate(option);
+    });
+    const docs = await query;
+
+    res.status(200).json({
+      status: 'success',
+      data: {
+        data: docs,
       },
     });
   });
